refactor: move components to appropriate folder
